Implemented smoother burger updates, bug fixes, test cases
- Changed burger updates to be shown to the user every certain amount of frames, changed burgers to be stored as floats to account for this
- Rewrote increment_burger_bps to account for this and round off any potential floating point rounding errors (determined by a float's decimal point being
larger than 0.9999)
- Fixed a bug where users could not purchase items from shop unless they had 1 more than the amount of burgers required
- Wrote test cases for the burger increment based on bps and truncate.

+++ src/config.ml
@@ -2,7 +2,7 @@ module R = Raylib
 (**[information] holds the user's stats — namely their burgers, bps, and clicking power.*)
 
 type information = {
-  mutable burgers : int;
+  mutable burgers : float;
   mutable bps : int;
   mutable click_power : int;
 }
